Apply the edit to the correct page
