+++ experiment/components/study/TaskPage.tsx
@@ -1,7 +1,7 @@
 'use client';
 
+import { useEffect, useRef, useState } from 'react';
 import { useSearchParams } from 'next/navigation';
-import { useRef } from 'react';
 import type { WritingAreaRef } from '@/components/WritingArea';
 import type { TextEditorState } from '@/types';
 import AIPanel from '@/components/AIPanel';
@@ -16,7 +16,20 @@ export default function TaskPage() {
   const username = searchParams.get('username') || '';
   const conditionCode = (searchParams.get('condition') || 'n') as keyof typeof letterToCondition; // TODO: don't default!
   const condition = letterToCondition[conditionCode];
-  const autoRefreshInterval = parseInt(searchParams.get('autoRefreshInterval') || '', 10) || 15000;
+
+  // Collapsible chat state
+  const [isChatOpen, setIsChatOpen] = useState(false);
+  const [hasUnread, setHasUnread] = useState(false);
+
+  // Auto-expand chat after a short delay so participants see Sarah's messages
+  useEffect(() => {
+    const delayTime = 1000; // 1 second
+    const timer = setTimeout(() => {
+      setIsChatOpen(true);
+    }, delayTime);
+
+    return () => clearTimeout(timer);
+  }, []);
 
   const handleSendTask = async (content: string) => {
     // Log task completion
@@ -49,37 +62,92 @@ export default function TaskPage() {
   };
 
   return (
-    <div className="flex flex-col h-screen bg-gray-100">
-      {/* Main content area */}
-      <div className="flex flex-1 overflow-hidden">
-        {/* Left column - Writing Area */}
-        <div className="flex-1 flex flex-col overflow-hidden bg-white">
-          <WritingArea
-            ref={writingAreaRef}
-            onSend={handleSendTask}
-            onUpdate={handleDocumentUpdate}
-            showSendButton={true}
-          />
-        </div>
+    <div className="flex h-screen gap-5 p-10 bg-gray-100 overflow-hidden relative">
+      {/* Left side - Writing Area with floating chat */}
+      <div className="relative flex-1 flex flex-col min-w-0">
+        <WritingArea
+          ref={writingAreaRef}
+          onSend={handleSendTask}
+          onUpdate={handleDocumentUpdate}
+          showSendButton={true}
+        />
 
-        {/* Right column - Chat and AI panels */}
-        <div className="w-96 flex flex-col border-l border-gray-300 overflow-hidden">
-          {/* Chat panel - top portion */}
-          <div className="flex-1 overflow-hidden border-b border-gray-300">
-            <ChatPanel />
-          </div>
+        {/* Collapsible Chat Window - floating over WritingArea */}
+        <div className="absolute bottom-0 right-8 z-50 flex flex-col items-end pointer-events-none">
+          <div
+            className={`rounded-t-lg bg-white border border-gray-300 pointer-events-auto flex flex-col transition-all duration-300 ease-in-out ${
+              isChatOpen ? 'w-[450px] h-[650px]' : 'w-[350px] h-12'
+            }`}
+          >
+            {/* Window Header (Tab) */}
+            <div
+              className={`h-12 flex items-center justify-between px-4 select-none transition-colors rounded-t-lg ${
+                isChatOpen ? 'bg-gray-50 border-b border-gray-300' : 'bg-white'
+              }`}
+            >
+              {isChatOpen ? (
+                <>
+                  <div className="flex items-center gap-2 font-semibold text-gray-700 text-sm">
+                    <span>Chat with Sarah</span>
+                  </div>
+                  <button
+                    type="button"
+                    onClick={() => setIsChatOpen(false)}
+                    className="p-1 hover:bg-gray-200 rounded cursor-pointer text-gray-500"
+                    aria-label="Minimize chat"
+                  >
+                    <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+                      <line x1="5" y1="12" x2="19" y2="12" />
+                    </svg>
+                  </button>
+                </>
+              ) : (
+                <button
+                  type="button"
+                  onClick={() => {
+                    setHasUnread(false);
+                    setIsChatOpen(true);
+                  }}
+                  className="flex-1 flex items-center justify-between hover:bg-gray-50 -mx-4 px-4 h-full cursor-pointer rounded-t-lg"
+                  aria-label="Open chat with Sarah"
+                >
+                  <div className="flex items-center gap-2 font-semibold text-gray-700 text-sm">
+                    <span>Chat with Sarah</span>
+                    {hasUnread && (
+                      <span className="flex h-5 w-5 items-center justify-center rounded-full bg-red-500 text-[10px] font-bold text-white shadow-sm">
+                        1
+                      </span>
+                    )}
+                  </div>
+                  <div className="flex items-center gap-1 text-gray-500">
+                    <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+                      <polyline points="18 15 12 9 6 15" />
+                    </svg>
+                  </div>
+                </button>
+              )}
+            </div>
 
-          {/* AI panel - bottom portion (only if not no_ai condition) */}
-          {condition !== 'no_ai' && (
-            <div className="flex-1 overflow-hidden">
-              <AIPanel
-                writingAreaRef={writingAreaRef}
-                isStudyMode={true}
+            {/* Window Body */}
+            <div className={`flex-1 overflow-hidden bg-white ${isChatOpen ? 'block' : 'hidden'}`}>
+              <ChatPanel
+                onNewMessage={() => {
+                  if (!isChatOpen) {
+                    setHasUnread(true);
+                  }
+                }}
               />
             </div>
-          )}
+          </div>
         </div>
       </div>
+
+      {/* Right side - AI Panel in sidebar (only for non-no_ai conditions) */}
+      {condition !== 'no_ai' && (
+        <div className="flex flex-col gap-2.5 w-110 border border-gray-300 rounded overflow-hidden shadow-sm bg-white">
+          <AIPanel writingAreaRef={writingAreaRef} isStudyMode={true} />
+        </div>
+      )}
     </div>
   );
 }
